refactor(api) migrate complementarycertification models
