+++ api/lib/domain/usecases/save-jury-complementary-certification-course-result.js
@@ -1,5 +1,5 @@
+import { ComplementaryCertificationCourseResult } from '../../../src/certification/complementary-certification/domain/models/ComplementaryCertificationCourseResult.js';
 import { InvalidJuryLevelError, NotFoundError } from '../errors.js';
-import { ComplementaryCertificationCourseResult } from '../models/ComplementaryCertificationCourseResult.js';
 
 const saveJuryComplementaryCertificationCourseResult = async function ({
   complementaryCertificationCourseId,
