+++ api/lib/infrastructure/repositories/complementary-certification-course-result-repository.js
@@ -1,5 +1,5 @@
 import { knex } from '../../../db/knex-database-connection.js';
-import { ComplementaryCertificationCourseResult } from '../../domain/models/ComplementaryCertificationCourseResult.js';
+import { ComplementaryCertificationCourseResult } from '../../../src/certification/complementary-certification/domain/models/ComplementaryCertificationCourseResult.js';
 
 const getPixSourceResultByComplementaryCertificationCourseId = async function ({ complementaryCertificationCourseId }) {
   const result = await knex
