+++ api/tests/acceptance/application/complementary-certification-course-results/complementary-certification-course-results-controller_test.js
@@ -1,4 +1,4 @@
-import { ComplementaryCertificationCourseResult } from '../../../../lib/domain/models/ComplementaryCertificationCourseResult.js';
+import { ComplementaryCertificationCourseResult } from '../../../../src/certification/complementary-certification/domain/models/ComplementaryCertificationCourseResult.js';
 import {
   createServer,
   databaseBuilder,
