+++ api/tests/integration/infrastructure/repositories/jury-certification-repository_test.js
@@ -1,6 +1,6 @@
 import { NotFoundError } from '../../../../lib/domain/errors.js';
-import { ComplementaryCertificationCourseResult } from '../../../../lib/domain/models/ComplementaryCertificationCourseResult.js';
 import * as juryCertificationRepository from '../../../../lib/infrastructure/repositories/jury-certification-repository.js';
+import { ComplementaryCertificationCourseResult } from '../../../../src/certification/complementary-certification/domain/models/ComplementaryCertificationCourseResult.js';
 import { AutoJuryCommentKeys } from '../../../../src/certification/shared/domain/models/JuryComment.js';
 import { catchErr, databaseBuilder, domainBuilder, expect } from '../../../test-helper.js';
