+++ api/tests/tooling/domain-builder/factory/build-complementary-certification-course-with-results.js
@@ -1,4 +1,4 @@
-import { ComplementaryCertificationCourseWithResults } from '../../../../lib/domain/models/ComplementaryCertificationCourseWithResults.js';
+import { ComplementaryCertificationCourseWithResults } from '../../../../src/certification/complementary-certification/domain/models/ComplementaryCertificationCourseWithResults.js';
 
 const buildComplementaryCertificationCourseWithResults = function ({
   id = 789,
