+++ api/tests/unit/domain/events/handle-complementary-certifications-scoring_test.js
@@ -2,7 +2,7 @@ import range from 'lodash/range.js';
 
 import { config as settings } from '../../../../lib/config.js';
 import { _forTestOnly } from '../../../../lib/domain/events/index.js';
-import { ComplementaryCertificationCourseResult } from '../../../../lib/domain/models/ComplementaryCertificationCourseResult.js';
+import { ComplementaryCertificationCourseResult } from '../../../../src/certification/complementary-certification/domain/models/ComplementaryCertificationCourseResult.js';
 import { status as assessmentResultStatuses } from '../../../../src/shared/domain/models/AssessmentResult.js';
 import { catchErr, domainBuilder, expect, sinon } from '../../../test-helper.js';
