+++ api/tests/unit/domain/read-models/CertifiedBadge_test.js
@@ -1,5 +1,5 @@
-import { ComplementaryCertificationCourseResult } from '../../../../lib/domain/models/ComplementaryCertificationCourseResult.js';
 import { CertifiedBadge } from '../../../../lib/domain/read-models/CertifiedBadge.js';
+import { ComplementaryCertificationCourseResult } from '../../../../src/certification/complementary-certification/domain/models/ComplementaryCertificationCourseResult.js';
 import { expect } from '../../../test-helper.js';
 
 describe('Unit | Domain | Read-models | CertifiedBadge', function () {
